// dashboard/src/legacy_pages/pages/Submissions/views/Sheet/types.ts
export interface Column<TRow, TSummaryRow = unknown> {
  /** The name of the column. By default it will be displayed in the header cell */
  name: string | ReactElement
  /** A unique key to distinguish each column */
  key: string
  /** Column left. */
  left?: number | string
  /** Column width. If not specified, it will be determined automatically based on grid width and specified widths of other columns */
  width?: number | string
  /** Minimum column width in px. */
  minWidth?: number
  /** Maximum column width in px. */
  maxWidth?: number
  cellClass?: string | ((row: TRow) => string | undefined)
  headerCellClass?: string
  summaryCellClass?: string | ((row: TSummaryRow) => string)
  /** Formatter to be used to render the cell content */
  formatter?: React.ComponentType<FormatterProps<TRow, TSummaryRow>>
  /** Formatter to be used to render the summary cell content */
  summaryFormatter?: React.ComponentType<SummaryFormatterProps<TSummaryRow, TRow>>
  /** Formatter to be used to render the group cell content */
  groupFormatter?: React.ComponentType<GroupFormatterProps<TRow, TSummaryRow>>
  /** Enables cell editing. If set and no editor property specified, then a textinput will be used as the cell editor */
  editable?: boolean | ((row: TRow) => boolean)
  /** Determines whether column is frozen or not */
  frozen?: boolean
  /** Enable resizing of a column */
  resizable?: boolean
  /** Enable sorting of a column */
  sortable?: boolean
  /** Sets the column sort order to be descending instead of ascending the first time the column is sorted */
  sortDescendingFirst?: boolean
  /** Editor to be rendered when cell of column is being edited. If set, then the column is automatically set to be editable */
  editor?: React.ComponentType<EditorProps<TRow, TSummaryRow>>
  editorOptions?: {
    /** @default false */
    createPortal?: boolean
    /** @default false */
    editOnClick?: boolean
    /** Prevent default to cancel editing */
    onCellKeyDown?: (event: React.KeyboardEvent<HTMLDivElement>) => void
    /** Control the default cell navigation behavior while the editor is open */
    onNavigation?: (event: React.KeyboardEvent<HTMLDivElement>) => boolean
    // TODO: Do we need these options
    // editOnDoubleClick?: boolean;
    /** @default false */
    // commitOnScroll?: boolean;
  }
  /** Header renderer for each header cell */
  headerRenderer?: React.ComponentType<HeaderRendererProps<TRow, TSummaryRow>>
  /** Component to be used to filter the data of the column */
  filterRenderer?: React.ComponentType<FilterRendererProps<TRow, any, TSummaryRow>>
}
// ...
export interface SheetColumn extends Column<any>, Pick<FormField, 'properties' | 'hide'> {
  idx?: number
  kind?: FieldKindEnum | InternalColumnKindEnum
}

export interface SheetKindIconProps extends ComponentProps {
  kind: FieldKindEnum | InternalColumnKindEnum
}

export interface ColorSelectorProps {
  color: string
  onChange: (color: string) => void
}

export interface ColumnOptions extends Partial<Pick<SheetColumn, 'kind' | 'name'>> {
  choices?: Choice[]
}
// ...
export type OnColumnOptionsUpdate = (column: SheetColumn, options: ColumnOptions) => void

// @ts-ignore
export interface SheetHeaderCellProps<R = any> extends HeaderRendererProps<R>, SharedCellProps {
  column: SheetColumn
  onColumnOptionsUpdate?: OnColumnOptionsUpdate
}

export interface SheetCellProps {
  rowIdx?: number
  column: SheetColumn
  row: Record<string, any>
  isCellSelected?: boolean
  onColumnOptionsUpdate?: OnColumnOptionsUpdate
  onCellValueChange?: (rowIdx: number, column: SheetColumn, value: any) => void
}